Add cross-validation, superoptimization, CITATION.cff
Cross-validation against independently written estimators. Everything else here
is self-consistent, which is not the same as correct: a systematic miscount
would show up identically on both sides of every internal comparison. Three
paths are used -- Google's Qualtran through Cirq, Qiskit's own accounting, and
a deliberately naive counter that re-derives everything from exported OpenQASM
3 *text* so it shares no data structure with the analyzer.

All three agree on qubits, Toffoli and Clifford counts for both the CDKM and
Gidney designs. Two things were fixed to get there, both real: Qualtran reports
a Clifford total rather than a CNOT count, so comparing it against our CNOTs
was a definitional mismatch and not a disagreement; and the QASM counter was
counting gate-*definition* bodies as invocations, inflating totals by exactly
the size of the custom AND gate definitions.

Superoptimization by meet-in-the-middle exhaustive synthesis. Depth-first
search is hopeless -- 28 gates on 4 qubits makes length 8 about 4x10^11 nodes --
but meeting in the middle needs two searches of length 4. The first meeting
point is a proof of optimality, since every shorter circuit was enumerated
first.

The result is more interesting than confirmation. The shortest reversible Ch is
3 gates using 2 Toffolis; the shortest Maj is 3 gates using 3. qSHA256 ships 4
gates with 1 Toffoli and 6 gates with 1. So the shortest circuit is decidedly
not the cheapest one: minimising gate count is the wrong objective under fault
tolerance, where Cliffords are nearly free and only the Toffolis are paid for.
That the exhaustive search and the algebraic multiplicative-complexity argument
independently agree on 1 Toffoli being achievable is the reassuring part.

CITATION.cff added so results are citable; version 2.2.0.

+++ qsha256/quantum/resources/layout.py
@@ -63,7 +63,7 @@ class LayoutModel:
     reference: str = "Litinski, 'A Game of Surface Codes', arXiv:1808.02892"
 
     def tile_count(self, logical_qubits: int) -> int:
-        return int(math.ceil(self.tiles(logical_qubits)))
+        return math.ceil(self.tiles(logical_qubits))
 
 
 LAYOUTS: dict[str, LayoutModel] = {
@@ -204,7 +204,7 @@ def lattice_surgery_layout(
 
     # A T gate every `cycles_per_t` cycles is what the floor plan permits; the
     # circuit's own T-depth is what it demands. The larger binds.
-    cycles_layout = int(math.ceil(t_count * layout.cycles_per_t))
+    cycles_layout = math.ceil(t_count * layout.cycles_per_t)
 
     distance = None
     cycles = max(1, cycles_layout)
